优化 复制文件时，先判定是否是文件，防止出现什么访问被拒绝的离题千里的错误

+++ BaseFileDirOperProject/FrmCopyFile.cs
@@ -75,6 +75,10 @@ namespace BaseFileDirOperProject
         }
         private void CopyFile(string src, string des)
         {
+            if(File.Exists(src))
+            {
+                throw new Exception($"抱歉，文件：{src}不存在！请检查！");
+            }
             File.Copy(src, des);
         }
